bigml-parallel: Factor out `arg=json` parsing & test it
